+++ backend/src/main/java/com/franbalsamo/mercadoapp/Model/DTO/VentaDTO.java
@@ -9,7 +9,7 @@ public class VentaDTO {
     private long id;
     private long id_boleta;
     private long id_producto;
-    private int cantidad;
+    private float cantidad;
     private float precio_unitario;
     private float precio_vacio;
